Extract factory methods to create instances of ArtifactSet into a separate class
