+++ subprojects/dependency-management/src/test/groovy/org/gradle/api/internal/artifacts/ivyservice/resolveengine/artifact/DefaultArtifactSetTest.groovy
@@ -57,9 +57,9 @@ class DefaultArtifactSetTest extends Specification {
         def ownerId = Stub(ModuleVersionIdentifier)
 
         given:
-        def artifacts1 = DefaultArtifactSet.createFromVariantMetadata(componentId, ownerId, null, null, [variant1, variant2] as Set, schema, null, null, artifactTypeRegistry, ImmutableAttributes.EMPTY, calculatedValueContainerFactory)
-        def artifacts2 = DefaultArtifactSet.createFromVariantMetadata(componentId, ownerId, null, null, [variant1] as Set, schema, null, null, artifactTypeRegistry, ImmutableAttributes.EMPTY, calculatedValueContainerFactory)
-        def artifacts3 = DefaultArtifactSet.adHocVariant(componentId, ownerId, [] as Set, null, null, schema, null, null, artifactTypeRegistry, ImmutableAttributes.EMPTY, ImmutableAttributes.EMPTY, calculatedValueContainerFactory)
+        def artifacts1 = ArtifactSetFactory.createFromVariantMetadata(componentId, ownerId, null, null, [variant1, variant2] as Set, schema, null, null, artifactTypeRegistry, ImmutableAttributes.EMPTY, calculatedValueContainerFactory)
+        def artifacts2 = ArtifactSetFactory.createFromVariantMetadata(componentId, ownerId, null, null, [variant1] as Set, schema, null, null, artifactTypeRegistry, ImmutableAttributes.EMPTY, calculatedValueContainerFactory)
+        def artifacts3 = ArtifactSetFactory.adHocVariant(componentId, ownerId, [] as Set, null, null, schema, null, null, artifactTypeRegistry, ImmutableAttributes.EMPTY, ImmutableAttributes.EMPTY, calculatedValueContainerFactory)
 
         ownerId.group >> "group"
         ownerId.name >> "name"
@@ -83,7 +83,7 @@ class DefaultArtifactSetTest extends Specification {
         def exclusions = new DefaultExcludeFactory().nothing()
 
         when:
-        def artifactSet = DefaultArtifactSet.createFromVariantMetadata(componentId, ownerId, moduleSources, exclusions, [variant1, variant2] as Set, schema, artifactResolver, new HashMap<ComponentArtifactIdentifier, ResolvableArtifact>(), artifactTypeRegistry, ImmutableAttributes.EMPTY, calculatedValueContainerFactory)
+        def artifactSet = ArtifactSetFactory.createFromVariantMetadata(componentId, ownerId, moduleSources, exclusions, [variant1, variant2] as Set, schema, artifactResolver, new HashMap<ComponentArtifactIdentifier, ResolvableArtifact>(), artifactTypeRegistry, ImmutableAttributes.EMPTY, calculatedValueContainerFactory)
         artifactSet.select({ true }, new VariantSelector() {
             @Override
             ResolvedArtifactSet select(ResolvedVariantSet candidates, VariantSelector.Factory factory) {
@@ -123,9 +123,9 @@ class DefaultArtifactSetTest extends Specification {
         def ownerId = Stub(ModuleVersionIdentifier)
 
         given:
-        def artifacts1 = DefaultArtifactSet.createFromVariantMetadata(componentId, ownerId, null, null, [variant1, variant2] as Set, schema, null, null, artifactTypeRegistry, ImmutableAttributes.EMPTY, calculatedValueContainerFactory)
-        def artifacts2 = DefaultArtifactSet.createFromVariantMetadata(componentId, ownerId, null, null, [variant1] as Set, schema, null, null, artifactTypeRegistry, ImmutableAttributes.EMPTY, calculatedValueContainerFactory)
-        def artifacts3 = DefaultArtifactSet.adHocVariant(componentId, ownerId, [] as Set, null, null, schema, null, null, artifactTypeRegistry, ImmutableAttributes.EMPTY, ImmutableAttributes.EMPTY, calculatedValueContainerFactory)
+        def artifacts1 = ArtifactSetFactory.createFromVariantMetadata(componentId, ownerId, null, null, [variant1, variant2] as Set, schema, null, null, artifactTypeRegistry, ImmutableAttributes.EMPTY, calculatedValueContainerFactory)
+        def artifacts2 = ArtifactSetFactory.createFromVariantMetadata(componentId, ownerId, null, null, [variant1] as Set, schema, null, null, artifactTypeRegistry, ImmutableAttributes.EMPTY, calculatedValueContainerFactory)
+        def artifacts3 = ArtifactSetFactory.adHocVariant(componentId, ownerId, [] as Set, null, null, schema, null, null, artifactTypeRegistry, ImmutableAttributes.EMPTY, ImmutableAttributes.EMPTY, calculatedValueContainerFactory)
 
         selector.select(_, _) >> resolvedVariant1
         ownerId.group >> "group"
